blam

+++ adapter_downloader.py
@@ -1,4 +1,5 @@
 import argparse
+import json
 import os
 import tempfile
 
@@ -6,6 +7,27 @@ import boto3
 import botocore
 import zipfile
 
+def get_adapter_info():
+
+    adapter_info = None
+    adapter_info_path = os.path.join(args.adapter_repo_path, args.adapter_info)
+
+    if os.path.exists(adapter_info_path):
+        f = open(adapter_info_path, "r")
+        adapter_info = json.loads(f.read())
+
+    if adapter_info == None:
+        adapter_info = dict()
+
+    return adapter_info
+
+def set_adapter_info(adapter_info):
+
+    adapter_info_path = os.path.join(args.adapter_repo_path, args.adapter_info)
+
+    json_object = json.dumps(adapter_info, indent=4)
+    with open(adapter_info_path, "w") as outfile:
+        outfile.write(json_object)
 
 if __name__ == '__main__':
 
@@ -19,6 +41,7 @@ if __name__ == '__main__':
     parser.add_argument('--aws_access_key_id', type=str, default='hQYiBAhIGNP5xIIU79yO', help='location of dataset')
     parser.add_argument('--aws_secret_access_key', type=str, default='jWNVPYT6zkxamILIG4YYIXGUQZkeJC39wJO2yQRb', help='location of dataset')
     parser.add_argument('--adapter_repo_path', type=str, default='adapters', help='location of dataset')
+    parser.add_argument('--adapter_info', type=str, default='adapter_info.json', help='location of dataset')
     # WARNING
 
     # get args
@@ -29,6 +52,9 @@ if __name__ == '__main__':
     if not isExist:
         os.makedirs(args.adapter_repo_path)
 
+    #get adapter info
+    adapter_info = get_adapter_info()
+
     s3 = boto3.resource('s3',
         endpoint_url=args.aws_endpoint_url,
         aws_access_key_id=args.aws_access_key_id,
@@ -38,26 +64,34 @@ if __name__ == '__main__':
     bucket = 'llmadapters'
     prefix = "llm_factory_trainer/"
 
+
     try:
 
-        my_bucket = s3.Bucket('llmadapters')
+        my_bucket = s3.Bucket(bucket)
         prefix = 'llm_factory_trainer/'
 
         for object in my_bucket.objects.filter(Prefix=prefix):
             if 'custom_adapter.zip' in object.key:
                 job_id = object.key.split('/')[1]
 
-                tmp_adapter_save_path = os.path.join(tempfile.gettempdir(), job_id + '.zip')
-                my_bucket.download_file(object.key, tmp_adapter_save_path)
+                if job_id not in adapter_info:
 
-                adapter_save_path = os.path.join(args.adapter_repo_path, job_id)
+                    print('New adapter for job_id:', job_id, 'found.')
+                    tmp_adapter_save_path = os.path.join(tempfile.gettempdir(), job_id + '.zip')
+                    print('Downloading: ', tmp_adapter_save_path)
+                    my_bucket.download_file(object.key, tmp_adapter_save_path)
+                    print('Saving adapter to repo')
+                    adapter_save_path = os.path.join(args.adapter_repo_path, job_id)
+                    with zipfile.ZipFile(tmp_adapter_save_path, 'r') as zip_ref:
+                        zip_ref.extractall(adapter_save_path)
+                    print('Removing temporary files:', tmp_adapter_save_path)
+                    #remove temp file
+                    os.remove(tmp_adapter_save_path)
 
-                with zipfile.ZipFile(tmp_adapter_save_path, 'r') as zip_ref:
-                    zip_ref.extractall(adapter_save_path)
+                    #record
+                    adapter_info[job_id] = adapter_save_path
 
-                #remove temp file
-                os.remove(tmp_adapter_save_path)
-
+        set_adapter_info(adapter_info)
 
     except botocore.exceptions.ClientError as e:
         if e.response['Error']['Code'] == "404":
